some various changes

+++ oak/src/oak/platform/OpenGL/OpenGLTexture.h
@@ -15,7 +15,7 @@ namespace Oak {
 
 		virtual uint32_t GetWidth() const override { return m_Width;  }
 		virtual uint32_t GetHeight() const override { return m_Height; }
-		virtual uint32_t GetRendererID() const override { return m_RendererID; }
+		virtual uint64_t GetRendererID() const override { return m_RendererID; }
 
 		virtual const std::string& GetPath() const override { return m_Path; }
