+++ oak/src/oak/renderer/Texture.h
@@ -13,7 +13,7 @@ namespace Oak {
 
 		virtual uint32_t GetWidth() const = 0;
 		virtual uint32_t GetHeight() const = 0;
-		virtual uint32_t GetRendererID() const = 0;
+		virtual uint64_t GetRendererID() const = 0;
 
 		virtual const std::string& GetPath() const = 0;
